+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## NEXT
+
+### Changed
+
+* Updated tests
+
 ## 0.2.0 - 2016-11-27
 
 ### Changed