+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+
+## Changed
+
+* Updated to `middlewares/utils#~0.9`
+* Improved route target resolution
+
 ## 0.3.0 - 2016-12-26
 
 ### Changed