+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.5.0 - 2017-02-27
+
+## Changed
+
+* Replaced `container-interop` by `psr/container`
+
 ## 0.4.0 - 2017-02-05
 
 ## Changed